start reviving memory debugging, part 3
The engine code depends on new mutable objects having their contents set to 0, which is
done by GC_MALLOC as part of the code for our operator new.  But that code is not run
when the object is allocated on the stack, so we have to track down such uses.